Rename VatPercentageField to NormalizedDecimalField
